+++ CHANGELOG.md
@@ -60,8 +60,8 @@
         * `onlySentMediaGroupUpdates`
         * `onlyEditMediaGroupUpdates`
     * Renames in chat filters extensions:
-        * `filterBaseMessageUpdates` -> `filterByChatId` and `filterByChat`
-        * `filterSentMediaGroupUpdates` -> `filterByChatId` and `filterByChat`
+        * `filterBaseMessageUpdates` -> `filterBaseMessageUpdatesByChatId` and `filterBaseMessageUpdatesByChat`
+        * `filterSentMediaGroupUpdates` -> `filterSentMediaGroupUpdatesByChatId` and `filterSentMediaGroupUpdatesByChat`
 
 ### 0.27.3